Refactor the type from Long to Integer in descricao in Categoria class

+++ Desafio-Modelo-De-Dominio-E-ORM/dominioeorm/src/main/java/com/desafio/dominioeorm/entities/Categoria.java
@@ -11,22 +11,22 @@ public class Categoria {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
     private String descricao;
 
     public Categoria() {
     }
 
-    public Categoria(Long id, String descricao) {
+    public Categoria(Integer id, String descricao) {
         this.id = id;
         this.descricao = descricao;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
